Carry Chest 1.0.0 (#1)
* Update CHANGELOG.md for version 3.2.1

- Added config option to reduce log verbosity.
- Enabled loading of config options from backup.
- Removed unnecessary harmony flag.
- Documented changes, additions, and fixes from previous versions.

* Update README.md for clarity and content adjustments

Removed "Credits" section and "Carry Chests" from the table of contents. Updated the link to the Generic Mod Config Menu and retained the Turkish translation entry in the translations list.

* Update version to stable release and improve comments

Updated the version number in `CarryChests.csproj` from
`1.0.0-beta.1` to `1.0.0` to reflect a stable release.
Additionally, modified a comment in `ModConfig.cs` for
better clarity and readability.